# Implement `compact_rep()`