--- a/src/slice-array.c
+++ b/src/slice-array.c
@@ -87,46 +87,99 @@ struct vec_slice_shaped_info {
   SEXP out_dim;
 };
 
-#define SLICE_SHAPED(RTYPE, CTYPE, DEREF, CONST_DEREF, NA_VALUE) \
-  SEXP out = PROTECT(Rf_allocArray(RTYPE, info.out_dim));        \
-  CTYPE* out_data = DEREF(out);                                  \
-  const CTYPE* x_data = CONST_DEREF(x);                          \
-                                                                 \
-  int out_loc = 0;                                               \
-                                                                 \
-  for (int i = 0; i < info.shape_elem_n; ++i) {                  \
-                                                                 \
-    /* Find and add the next `x` element */                      \
-    for (int j = 0; j < info.index_n; ++j) {                     \
-      int size_index = info.p_index[j];                          \
-                                                                 \
-      if (size_index == NA_INTEGER) {                            \
-        out_data[out_loc] = NA_VALUE;                            \
-      } else {                                                   \
-        int loc = vec_strided_loc(                               \
-          size_index - 1,                                        \
-          info.p_shape_index,                                    \
-          info.p_strides,                                        \
-          info.shape_n                                           \
-        );                                                       \
-        out_data[out_loc] = x_data[loc];                         \
-      }                                                          \
-                                                                 \
-      out_loc++;                                                 \
-    }                                                            \
-                                                                 \
-    /* Update shape_index */                                     \
-    for (int j = 0; j < info.shape_n; ++j) {                     \
-      info.p_shape_index[j]++;                                   \
-      if (info.p_shape_index[j] < info.p_dim[j + 1]) {           \
-        break;                                                   \
-      }                                                          \
-      info.p_shape_index[j] = 0;                                 \
-    }                                                            \
-  }                                                              \
-                                                                 \
-  UNPROTECT(1);                                                  \
+#define SLICE_SHAPED_INDEX(RTYPE, CTYPE, DEREF, CONST_DEREF, NA_VALUE) \
+  SEXP out = PROTECT(Rf_allocArray(RTYPE, info.out_dim));              \
+  CTYPE* out_data = DEREF(out);                                        \
+  const CTYPE* x_data = CONST_DEREF(x);                                \
+                                                                       \
+  int out_loc = 0;                                                     \
+                                                                       \
+  for (int i = 0; i < info.shape_elem_n; ++i) {                        \
+                                                                       \
+    /* Find and add the next `x` element */                            \
+    for (int j = 0; j < info.index_n; ++j) {                           \
+      int size_index = info.p_index[j];                                \
+                                                                       \
+      if (size_index == NA_INTEGER) {                                  \
+        out_data[out_loc] = NA_VALUE;                                  \
+      } else {                                                         \
+        int loc = vec_strided_loc(                                     \
+          size_index - 1,                                              \
+          info.p_shape_index,                                          \
+          info.p_strides,                                              \
+          info.shape_n                                                 \
+        );                                                             \
+        out_data[out_loc] = x_data[loc];                               \
+      }                                                                \
+                                                                       \
+      out_loc++;                                                       \
+    }                                                                  \
+                                                                       \
+    /* Update shape_index */                                           \
+    for (int j = 0; j < info.shape_n; ++j) {                           \
+      info.p_shape_index[j]++;                                         \
+      if (info.p_shape_index[j] < info.p_dim[j + 1]) {                 \
+        break;                                                         \
+      }                                                                \
+      info.p_shape_index[j] = 0;                                       \
+    }                                                                  \
+  }                                                                    \
+                                                                       \
+  UNPROTECT(1);                                                        \
   return out
+
+#define SLICE_SHAPED_COMPACT_REP(RTYPE, CTYPE, DEREF, CONST_DEREF, NA_VALUE)   \
+  SEXP out = PROTECT(Rf_allocArray(RTYPE, info.out_dim));                      \
+  CTYPE* out_data = DEREF(out);                                                \
+                                                                               \
+  int size_index = info.p_index[0];                                            \
+  if (size_index == NA_INTEGER) {                                              \
+    for (int i = 0; i < info.shape_elem_n; ++i, ++out_data) {                  \
+      *out_data = NA_VALUE;                                                    \
+    }                                                                          \
+    UNPROTECT(1);                                                              \
+    return(out);                                                               \
+  }                                                                            \
+                                                                               \
+  const CTYPE* x_data = CONST_DEREF(x);                                        \
+  int out_loc = 0;                                                             \
+                                                                               \
+  /* Convert to C index */                                                     \
+  size_index = size_index - 1;                                                 \
+                                                                               \
+  for (int i = 0; i < info.shape_elem_n; ++i) {                                \
+                                                                               \
+    /* Find and add the next `x` element */                                    \
+    for (int j = 0; j < info.index_n; ++j) {                                   \
+      int loc = vec_strided_loc(                                               \
+        size_index,                                                            \
+        info.p_shape_index,                                                    \
+        info.p_strides,                                                        \
+        info.shape_n                                                           \
+      );                                                                       \
+      out_data[out_loc] = x_data[loc];                                         \
+      out_loc++;                                                               \
+    }                                                                          \
+                                                                               \
+    /* Update shape_index */                                                   \
+    for (int j = 0; j < info.shape_n; ++j) {                                   \
+      info.p_shape_index[j]++;                                                 \
+      if (info.p_shape_index[j] < info.p_dim[j + 1]) {                         \
+        break;                                                                 \
+      }                                                                        \
+      info.p_shape_index[j] = 0;                                               \
+    }                                                                          \
+  }                                                                            \
+                                                                               \
+  UNPROTECT(1);                                                                \
+  return out
+
+#define SLICE_SHAPED(RTYPE, CTYPE, DEREF, CONST_DEREF, NA_VALUE)          \
+  if (is_compact_rep(index)) {                                            \
+    SLICE_SHAPED_COMPACT_REP(RTYPE, CTYPE, DEREF, CONST_DEREF, NA_VALUE); \
+  } else {                                                                \
+    SLICE_SHAPED_INDEX(RTYPE, CTYPE, DEREF, CONST_DEREF, NA_VALUE);       \
+  }
 
 static SEXP lgl_slice_shaped(SEXP x, SEXP index, struct vec_slice_shaped_info info) {
   SLICE_SHAPED(LGLSXP, int, LOGICAL, LOGICAL_RO, NA_LOGICAL);
@@ -148,8 +201,10 @@ static SEXP raw_slice_shaped(SEXP x, SEXP index, struct vec_slice_shaped_info in
 }
 
 #undef SLICE_SHAPED
+#undef SLICE_SHAPED_COMPACT_REP
+#undef SLICE_SHAPED_INDEX
 
-#define SLICE_BARRIER_SHAPED(RTYPE, GET, SET, NA_VALUE)        \
+#define SLICE_BARRIER_SHAPED_INDEX(RTYPE, GET, SET, NA_VALUE)  \
   SEXP out = PROTECT(Rf_allocArray(RTYPE, info.out_dim));      \
                                                                \
   int out_loc = 0;                                             \
@@ -189,11 +244,65 @@ static SEXP raw_slice_shaped(SEXP x, SEXP index, struct vec_slice_shaped_info in
   UNPROTECT(1);                                                \
   return out
 
+#define SLICE_BARRIER_SHAPED_COMPACT_REP(RTYPE, GET, SET, NA_VALUE)   \
+  SEXP out = PROTECT(Rf_allocArray(RTYPE, info.out_dim));             \
+                                                                      \
+  int size_index = info.p_index[0];                                   \
+  if (size_index == NA_INTEGER) {                                     \
+    for (int i = 0; i < info.shape_elem_n; ++i) {                     \
+      SET(out, i, NA_VALUE);                                          \
+    }                                                                 \
+    UNPROTECT(1);                                                     \
+    return(out);                                                      \
+  }                                                                   \
+                                                                      \
+  int out_loc = 0;                                                    \
+                                                                      \
+  /* Convert to C index */                                            \
+  size_index = size_index - 1;                                        \
+                                                                      \
+  for (int i = 0; i < info.shape_elem_n; ++i) {                       \
+                                                                      \
+    /* Find and add the next `x` element */                           \
+    for (int j = 0; j < info.index_n; ++j) {                          \
+      int loc = vec_strided_loc(                                      \
+        size_index,                                                   \
+        info.p_shape_index,                                           \
+        info.p_strides,                                               \
+        info.shape_n                                                  \
+      );                                                              \
+      SEXP elt = GET(x, loc);                                         \
+      SET(out, out_loc, elt);                                         \
+      out_loc++;                                                      \
+    }                                                                 \
+                                                                      \
+    /* Update shape_index */                                          \
+    for (int j = 0; j < info.shape_n; ++j) {                          \
+      info.p_shape_index[j]++;                                        \
+      if (info.p_shape_index[j] < info.p_dim[j + 1]) {                \
+        break;                                                        \
+      }                                                               \
+      info.p_shape_index[j] = 0;                                      \
+    }                                                                 \
+  }                                                                   \
+                                                                      \
+  UNPROTECT(1);                                                       \
+  return out
+
+#define SLICE_BARRIER_SHAPED(RTYPE, GET, SET, NA_VALUE)          \
+  if (is_compact_rep(index)) {                                   \
+    SLICE_BARRIER_SHAPED_COMPACT_REP(RTYPE, GET, SET, NA_VALUE); \
+  } else {                                                       \
+    SLICE_BARRIER_SHAPED_INDEX(RTYPE, GET, SET, NA_VALUE);       \
+  }
+
 static SEXP list_slice_shaped(SEXP x, SEXP index, struct vec_slice_shaped_info info) {
   SLICE_BARRIER_SHAPED(VECSXP, VECTOR_ELT, SET_VECTOR_ELT, R_NilValue);
 }
 
 #undef SLICE_BARRIER_SHAPED
+#undef SLICE_BARRIER_SHAPED_COMPACT_REP
+#undef SLICE_BARRIER_SHAPED_INDEX
 
 SEXP vec_slice_shaped_base(enum vctrs_type type,
                            SEXP x,
@@ -221,7 +330,7 @@ SEXP vec_slice_shaped(enum vctrs_type type, SEXP x, SEXP index) {
   info.p_index = INTEGER_RO(index);
   info.dim_n = Rf_length(dim);
   info.shape_n = info.dim_n - 1;
-  info.index_n = Rf_length(index);
+  info.index_n = is_compact_rep(index) ? info.p_index[1] : Rf_length(index);
 
   SEXP strides = PROTECT(vec_strides(info.p_dim, info.shape_n));
   info.p_strides = INTEGER_RO(strides);