--- a/src/utils.h
+++ b/src/utils.h
@@ -81,6 +81,7 @@ bool is_compact_seq(SEXP x);
 void init_compact_rep(int* p, R_len_t i, R_len_t n);
 SEXP compact_rep(R_len_t i, R_len_t n);
 bool is_compact_rep(SEXP x);
+SEXP compact_rep_materialize(SEXP x);
 
 SEXP apply_name_spec(SEXP name_spec, SEXP outer, SEXP inner, R_len_t n);
 SEXP outer_names(SEXP names, SEXP outer, R_len_t n);