Closes #472 

This PR introduces `compact_rep()`. It is similar to `compact_seq()`, but takes a single value `i` and repeats it `n` times.

**With this PR we gain performance at the cost of more C defines**

`vec_slice()` has been made aware of `compact_rep` objects, and `vec_recycle()` and the C level `vec_init()` use it.

Because `vec_init()` and `vec_recycle()` are used in `vec_c()`, `vec_rbind()`, and `vec_cbind()`, these all inherit a little bit of the performance.

_It would be great if the R level `vec_init()` used the C level version of `vec_init()` so it could get these benefits too._

``` r
library(vctrs)
library(bench)
```

``` r
# old (recycling length 1 object)
mark(recycle = vec_recycle(1L, 10000L), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 recycle      14.4µs     23µs    41152.    80.6KB     56.0

# new (recycling length 1 object)
mark(recycle = vec_recycle(1L, 10000L), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 recycle       2.8µs   6.02µs   136939.    45.6KB     93.2
```

``` r
x <- 1:1000

# old (vec_init()'s the output container)
mark(c = vec_c(x, x, x, x, x), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 c            9.29µs     13µs    66621.    46.2KB     48.0

# new (vec_init()'s the output container)
mark(c = vec_c(x, x, x, x, x), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 c            5.39µs   7.71µs   116337.    30.5KB     40.7
```

``` r
x <- data.frame(x = 1:1000)

# old (vec_init()'s the output container)
mark(rbind = vec_rbind(x,x,x,x,x), iterations = 10000)
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 rbind        53.6µs   68.2µs    14374.    67.9KB     15.8

# new (vec_init()'s the output container)
mark(rbind = vec_rbind(x,x,x,x,x), iterations = 10000)
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 rbind        48.7µs   59.9µs    16786.    53.8KB     15.1
```

``` r
x <- data.frame(x = 1:100000)

# old (vec_init()'s the output container)
mark(rbind_big = vec_rbind(x,x,x,x,x), iterations = 10000)
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 rbind_big    2.38ms   3.39ms      289.     6.1MB     31.4

# new (vec_init()'s the output container)
mark(rbind_big = vec_rbind(x,x,x,x,x), iterations = 10000)
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 rbind_big     2.2ms   3.02ms      325.     4.2MB     22.4
```

``` r
x <- data.frame(x = 1:100000)
y <- data.frame(y = 1)

# old (vec_recycle()'s size 1 input)
mark(cbind = vec_cbind(x,y,y,y,y, .name_repair = "minimal"), iterations = 10000)
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 cbind         650µs    1.1ms      900.    4.58MB     120.

# new (vec_recycle()'s size 1 input)
mark(cbind = vec_cbind(x,y,y,y,y, .name_repair = "minimal"), iterations = 10000)
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 cbind         248µs    441µs     2253.    3.06MB     163.
```