@lionel- 2019-07-02
> It would be great if the R level vec_init() used the C level version of vec_init() so it could get these benefits too.

This is planned with ALTREP vectors. Ideally we'd generate the same kind of altrep vectors as `base::seq()` and `base::rep()` but I haven't looked into whether that's possible yet.