@DavisVaughan 2019-07-02
> For completeness we're now missing compact seqs in `slice()` and compact reps in `assign()` right?

Correct! I can't think of a use case for compact reps in `assign()`, and compact seqs in `slice()` haven't come up yet. But yes for completeness that is missing.