@lionel- 2019-07-02
Agreed there is no need to implement them now, we can tackle this as we implement support for altrep seqs and reps.