@DavisVaughan 2019-07-02
It's a bit hard to add tests specifically for this case as they aren't really exposed to the R side, and their usage in the C `vec_init()` and `vec_recycle()` don't cover everything. I found a small bug here with `NA` indices that I fixed. https://github.com/r-lib/vctrs/pull/473/commits/4b9acb0cb69ad4b01b3a8b5d3ed4786fde7199f5

I had to expose the C `vec_init()` for a minute to be able to test it but it is working correctly now. We'd have to fully expose the compact rep idea to the R side AND make the R version of `vec_slice()` aware of it to fully test. Otherwise as shown in the above examples you can see it working implicitly, but those wouldn't make good tests.