@lionel- 2019-07-02
It would make sense to expose the creation of compact reps to the R side just for tests. But if it's too much work to make `vec_slice()` compatible with compact vectors, maybe it's better to wait for the altrep implementation.