@DavisVaughan 2019-07-02
I think it just requires making `vec_as_index()` aware of compact reps. I'll give it a try